fix: address code review findings across actor-core
- docs: add blank lines around headings/tables (MD022/MD058)
- refactor(receptionist): remove redundant second lookup in get()
- fix(behavior_runner): capture has_signal_handler before dispatch_signal
- refactor(worker_stats): change worker count type from u32 to usize
- style(tests): replace English Given/When/Then comments with Japanese rationale
- refactor(tests): add use alias for long WorkPullingProducerControllerCommandKind path
- fix(kernel_public_surface): add visibility diagnostic assertion, fix TOCTOU in unique_crate_dir

+++ modules/actor-core/src/core/typed/delivery/internal/work_pulling_producer_controller.rs
@@ -177,8 +177,8 @@ where
   }
 
   /// Returns the number of registered workers.
-  fn worker_count(&self) -> u32 {
-    self.workers.len() as u32
+  fn worker_count(&self) -> usize {
+    self.workers.len()
   }
 
   /// Finds a worker key that has demand, using round-robin starting from
